FlightScheduleController + repos take integer instead of long
@GetMapping
@Post
@Put
@Delete

+++ FRS/src/main/java/com/group/FRS/repository/PassengerRepository.java
@@ -7,7 +7,7 @@ import com.group.FRS.model.Passenger;
 
 
 @Repository
-public interface PassengerRepository extends JpaRepository<Passenger, Long> {
+public interface PassengerRepository extends JpaRepository<Passenger, Integer> {
 //
 }
