+++ FRS/src/main/java/com/group/FRS/repository/User_CredentialRepository.java
@@ -7,6 +7,6 @@ import com.group.FRS.model.UserCredential;
 
 
 @Repository
-public interface User_CredentialRepository extends JpaRepository<UserCredential, Long> {
+public interface User_CredentialRepository extends JpaRepository<UserCredential, Integer> {
 	//
 	}
